requirements.txt
matplotlib>=3.0.0,<4.0.0
numpy>=2.0.0,<3.0.0
pandas>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.0.0,<3.0.0